Adding tests and fixing issues in the backend

+++ PhoneBook/phonebook-api/src/main/java/com/livebox/phonebookapi/service/SingleParamLookupStrategy.java
@@ -22,7 +22,7 @@ public class SingleParamLookupStrategy extends ContactLookupStrategy{
         final String stringParam = (String) param;
 
         try{
-            int phoneNumber = Integer.parseInt(stringParam);
+            long phoneNumber = Long.parseLong(stringParam);
             return phoneBookRepository.filterContactsByMatchingNumber(phoneNumber);
         }catch(NumberFormatException numberFormatException){
             return phoneBookRepository.filterContactsByMatchingName(stringParam.toLowerCase());
